Update - NumOccurences & array C size
I added a portion that not only set array size C to be the range of A
NOT range of A starting at zero for array index purposes... the array
starts at lowest number.

NumOccurences:
Make C(counter): length be range of A
size: 0-5 = 6
Count how many of that certain number exist.
ex: how many 3's?
know what index u are in, add the occurences of that index number in A
into C